pyproject.toml
[project]
name = "easypos"
version = "0.1.0"
description = ""
authors = [
    {name = "Samuel Silva",email = "samuelsilva2310@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"

dependencies = [
    "python-escpos[all] (>=3.1,<4.0)",
    "PySide6 (>=6.9.2)",
    "customtkinter (>=5.2.2,<6.0.0)",
]

[tool.poetry]
packages = [{include = "easypos", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
